requirements.txt
flask>=2.0.0
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.0.0
pandas>=1.3.0
open-clip-torch>=2.0.0
werkzeug>=2.0.0
scikit-learn>=1.0.0